Issue #625 Extract run scheduling functionality to its own components.
